my-ai-app/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sse-starlette
google-generativeai